Perf/Move/Remove/ interest 폴더 엔티티, dto 옮김 및 연동 #15
